move extension to EfExtension class
